// frontend/src/api/client.ts
export class CreateCidadeDto implements ICreateCidadeDto {
    cidade?: string;
    ddd?: number;
    estadoId?: number;

    constructor(data?: ICreateCidadeDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.cidade = _data["cidade"];
            this.ddd = _data["ddd"];
            this.estadoId = _data["estadoId"];
        }
    }

    static fromJS(data: any): CreateCidadeDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreateCidadeDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["cidade"] = this.cidade;
        data["ddd"] = this.ddd;
        data["estadoId"] = this.estadoId;
        return data;
    }
}

export interface ICreateCidadeDto {
    cidade?: string;
    ddd?: number;
    estadoId?: number;
}

export class UpdateCidadeDto implements IUpdateCidadeDto {
    cidade?: string;
    ddd?: number;
    estadoId?: number;

    constructor(data?: IUpdateCidadeDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.cidade = _data["cidade"];
            this.ddd = _data["ddd"];
            this.estadoId = _data["estadoId"];
        }
    }

    static fromJS(data: any): UpdateCidadeDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateCidadeDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["cidade"] = this.cidade;
        data["ddd"] = this.ddd;
        data["estadoId"] = this.estadoId;
        return data;
    }
}

export interface IUpdateCidadeDto {
    cidade?: string;
    ddd?: number;
    estadoId?: number;
}